+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the library will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2.1.1
+> 2021-05-14
+
+- Renamed ToggleSwitchViewModel's property `isEnabled` to `enabled` to fix name conflict when combined with the iOS lib of `trikot.analytics`
+- Renamed ToggleSwitchViewModel's property `isOn` to `checked` to follow naming convention
+
+
 ## 2.1.0
 > 2021-05-10